fix(jest): Fix watchmode on api tests by loading scenarios in the same context as jest
